\name{ex01.25}
\alias{ex01.25}
\docType{data}
\title{R Data set: ex01.25}
\description{
The \code{ex01.25} data frame has 40 rows and 2 columns.
}
\usage{data(ex01.25)}
\format{
  A data frame with 40 observations on the following 2 variables.
  \describe{
    \item{\code{IDT}}{a numeric vector}
    \item{\code{log10.IDT}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex01.25)
str(ex01.25)
}
\keyword{datasets}
