\name{ex01.28}
\alias{ex01.28}
\docType{data}
\title{R Data set: ex01.28}
\description{
The \code{ex01.28} data frame has 60 rows and 1 column.
}
\usage{data(ex01.28)}
\format{
  A data frame with 60 observations on the following variable.
  \describe{
    \item{\code{radiation}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex01.28)
str(ex01.28)
}
\keyword{datasets}
