\name{ex01.29}
\alias{ex01.29}
\docType{data}
\title{R Data set: ex01.29}
\description{
The \code{ex01.29} data frame has 61 rows and 1 column.
}
\usage{data(ex01.29)}
\format{
  A data frame with 61 observations on the following variable.
  \describe{
    \item{\code{C1}}{a factor with levels \code{B} \code{C} \code{C5} \code{F} \code{J} \code{M} \code{N} \code{O}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex01.29)
str(ex01.29)
}
\keyword{datasets}
