\name{ex01.32}
\alias{ex01.32}
\docType{data}
\title{R Data set: ex01.32}
\description{
The \code{ex01.32} data frame has 14 rows and 2 columns.
}
\usage{data(ex01.32)}
\format{
  A data frame with 14 observations on the following 2 variables.
  \describe{
    \item{\code{Value}}{a numeric vector}
    \item{\code{Cumulative}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex01.32)
str(ex01.32)
}
\keyword{datasets}
