\name{ex01.60}
\alias{ex01.60}
\docType{data}
\title{R Data set: ex01.60}
\description{
The \code{ex01.60} data frame has 12 rows and 2 columns.
}
\usage{data(ex01.60)}
\format{
  A data frame with 12 observations on the following 2 variables.
  \describe{
    \item{\code{Test}}{a numeric vector}
    \item{\code{Cannister}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex01.60)
str(ex01.60)
}
\keyword{datasets}
