\name{ex01.70}
\alias{ex01.70}
\docType{data}
\title{R Data set: ex01.70}
\description{
The \code{ex01.70} data frame has 15 rows and 2 columns.
}
\usage{data(ex01.70)}
\format{
  A data frame with 15 observations on the following 2 variables.
  \describe{
    \item{\code{Weight}}{a numeric vector}
    \item{\code{Treadmill}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex01.70)
str(ex01.70)
}
\keyword{datasets}
