\name{ex01.75}
\alias{ex01.75}
\docType{data}
\title{R Data set: ex01.75}
\description{
The \code{ex01.75} data frame has 15 rows and 3 columns.
}
\usage{data(ex01.75)}
\format{
  A data frame with 15 observations on the following 3 variables.
  \describe{
    \item{\code{Type.1}}{a numeric vector}
    \item{\code{Type.2}}{a numeric vector}
    \item{\code{Type.3}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex01.75)
str(ex01.75)
}
\keyword{datasets}
