\name{ex04.83}
\alias{ex04.83}
\docType{data}
\title{R Data set: ex04.83}
\description{
The \code{ex04.83} data frame has 16 rows and 1 column.
}
\usage{data(ex04.83)}
\format{
  A data frame with 16 observations on the following variable.
  \describe{
    \item{\code{thickness}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex04.83)
str(ex04.83)
}
\keyword{datasets}
