\name{ex04.90}
\alias{ex04.90}
\docType{data}
\title{R Data set: ex04.90}
\description{
The \code{ex04.90} data frame has 18 rows and 2 columns.
}
\usage{data(ex04.90)}
\format{
  A data frame with 18 observations on the following 2 variables.
  \describe{
    \item{\code{obsv}}{a numeric vector}
    \item{\code{p}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex04.90)
str(ex04.90)
}
\keyword{datasets}
