\name{ex06.02}
\alias{ex06.02}
\docType{data}
\title{R Data set: ex06.02}
\description{
The \code{ex06.02} data frame has 21 rows and 1 column.
}
\usage{data(ex06.02)}
\format{
  A data frame with 21 observations on the following variable.
  \describe{
    \item{\code{C1}}{a factor with levels \code{C} \code{C1} \code{H} \code{S} \code{T}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex06.02)
str(ex06.02)
}
\keyword{datasets}
