\name{ex06.04}
\alias{ex06.04}
\docType{data}
\title{R Data set: ex06.04}
\description{
The \code{ex06.04} data frame has 20 rows and 1 column.
}
\usage{data(ex06.04)}
\format{
  A data frame with 20 observations on the following variable.
  \describe{
    \item{\code{C1}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex06.04)
str(ex06.04)
}
\keyword{datasets}
