\name{ex06.05}
\alias{ex06.05}
\docType{data}
\title{R Data set: ex06.05}
\description{
The \code{ex06.05} data frame has 5 rows and 3 columns.
}
\usage{data(ex06.05)}
\format{
  A data frame with 5 observations on the following 3 variables.
  \describe{
    \item{\code{Book.value}}{a numeric vector}
    \item{\code{Audited.value}}{a numeric vector}
    \item{\code{Error}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex06.05)
str(ex06.05)
}
\keyword{datasets}
