\name{ex06.09}
\alias{ex06.09}
\docType{data}
\title{R Data set: ex06.09}
\description{
The \code{ex06.09} data frame has 8 rows and 2 columns.
}
\usage{data(ex06.09)}
\format{
  A data frame with 8 observations on the following 2 variables.
  \describe{
    \item{\code{Number.of.searches.per.item}}{a numeric vector}
    \item{\code{Observed.frequency}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex06.09)
str(ex06.09)
}
\keyword{datasets}
