\name{ex07.26}
\alias{ex07.26}
\docType{data}
\title{R Data set: ex07.26}
\description{
The \code{ex07.26} data frame has 11 rows and 2 columns.
}
\usage{data(ex07.26)}
\format{
  A data frame with 11 observations on the following 2 variables.
  \describe{
    \item{\code{Number.of.absences}}{a numeric vector}
    \item{\code{Frequency}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex07.26)
str(ex07.26)
}
\keyword{datasets}
