\name{ex07.45}
\alias{ex07.45}
\docType{data}
\title{R Data set: ex07.45}
\description{
The \code{ex07.45} data frame has 22 rows and 1 column.
}
\usage{data(ex07.45)}
\format{
  A data frame with 22 observations on the following variable.
  \describe{
    \item{\code{C1}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex07.45)
str(ex07.45)
}
\keyword{datasets}
