\name{ex08.56}
\alias{ex08.56}
\docType{data}
\title{R Data set: ex08.56}
\description{
The \code{ex08.56} data frame has 30 rows and 1 column.
}
\usage{data(ex08.56)}
\format{
  A data frame with 30 observations on the following variable.
  \describe{
    \item{\code{percorg}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex08.56)
str(ex08.56)
}
\keyword{datasets}
