\name{ex09.07}
\alias{ex09.07}
\docType{data}
\title{R Data set: ex09.07}
\description{
The \code{ex09.07} data frame has 2 rows and 4 columns.
}
\usage{data(ex09.07)}
\format{
  A data frame with 2 observations on the following 4 variables.
  \describe{
    \item{\code{Gender}}{a factor with levels \code{Females} \code{Males}}
    \item{\code{Sample.Size}}{a numeric vector}
    \item{\code{Sample.Mean}}{a numeric vector}
    \item{\code{Sample.Standard.Deviation}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex09.07)
str(ex09.07)
}
\keyword{datasets}
