\name{ex09.16}
\alias{ex09.16}
\docType{data}
\title{R Data set: ex09.16}
\description{
The \code{ex09.16} data frame has 2 rows and 3 columns.
}
\usage{data(ex09.16)}
\format{
  A data frame with 2 observations on the following 3 variables.
  \describe{
    \item{\code{Type}}{a numeric vector}
    \item{\code{Sample.Average}}{a numeric vector}
    \item{\code{Sample.Standard.Deviation}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex09.16)
str(ex09.16)
}
\keyword{datasets}
