\name{ex09.27}
\alias{ex09.27}
\docType{data}
\title{R Data set: ex09.27}
\description{
The \code{ex09.27} data frame has 2 rows and 4 columns.
}
\usage{data(ex09.27)}
\format{
  A data frame with 2 observations on the following 4 variables.
  \describe{
    \item{\code{Type.of.Player}}{a factor with levels \code{Advanced} \code{Intermediate}}
    \item{\code{Sample.size}}{a numeric vector}
    \item{\code{Sample.mean}}{a numeric vector}
    \item{\code{Sample.standard.deviation}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex09.27)
str(ex09.27)
}
\keyword{datasets}
