\name{ex09.37}
\alias{ex09.37}
\docType{data}
\title{R Data set: ex09.37}
\description{
The \code{ex09.37} data frame has 33 rows and 3 columns.
}
\usage{data(ex09.37)}
\format{
  A data frame with 33 observations on the following 3 variables.
  \describe{
    \item{\code{House}}{a numeric vector}
    \item{\code{Indoor}}{a numeric vector}
    \item{\code{Outdoor}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex09.37)
str(ex09.37)
}
\keyword{datasets}
