\name{ex09.38}
\alias{ex09.38}
\docType{data}
\title{R Data set: ex09.38}
\description{
The \code{ex09.38} data frame has 15 rows and 3 columns.
}
\usage{data(ex09.38)}
\format{
  A data frame with 15 observations on the following 3 variables.
  \describe{
    \item{\code{Test.condition}}{a numeric vector}
    \item{\code{Normal}}{a numeric vector}
    \item{\code{High}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex09.38)
str(ex09.38)
}
\keyword{datasets}
