\name{ex09.39}
\alias{ex09.39}
\docType{data}
\title{R Data set: ex09.39}
\description{
The \code{ex09.39} data frame has 14 rows and 4 columns.
}
\usage{data(ex09.39)}
\format{
  A data frame with 14 observations on the following 4 variables.
  \describe{
    \item{\code{Infant}}{a numeric vector}
    \item{\code{Isotopic.method}}{a numeric vector}
    \item{\code{Test}}{a numeric vector}
    \item{\code{Difference}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex09.39)
str(ex09.39)
}
\keyword{datasets}
