\name{ex09.40}
\alias{ex09.40}
\docType{data}
\title{R Data set: ex09.40}
\description{
The \code{ex09.40} data frame has 16 rows and 3 columns.
}
\usage{data(ex09.40)}
\format{
  A data frame with 16 observations on the following 3 variables.
  \describe{
    \item{\code{Period}}{a numeric vector}
    \item{\code{Pipe}}{a numeric vector}
    \item{\code{Brush}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex09.40)
str(ex09.40)
}
\keyword{datasets}
