\name{ex09.41}
\alias{ex09.41}
\docType{data}
\title{R Data set: ex09.41}
\description{
The \code{ex09.41} data frame has 9 rows and 3 columns.
}
\usage{data(ex09.41)}
\format{
  A data frame with 9 observations on the following 3 variables.
  \describe{
    \item{\code{Subject}}{a numeric vector}
    \item{\code{Black}}{a numeric vector}
    \item{\code{White}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex09.41)
str(ex09.41)
}
\keyword{datasets}
