\name{ex09.43}
\alias{ex09.43}
\docType{data}
\title{R Data set: ex09.43}
\description{
The \code{ex09.43} data frame has 15 rows and 1 column.
}
\usage{data(ex09.43)}
\format{
  A data frame with 15 observations on the following variable.
  \describe{
    \item{\code{c1}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex09.43)
str(ex09.43)
}
\keyword{datasets}
