\name{ex09.44}
\alias{ex09.44}
\docType{data}
\title{R Data set: ex09.44}
\description{
The \code{ex09.44} data frame has 16 rows and 2 columns.
}
\usage{data(ex09.44)}
\format{
  A data frame with 16 observations on the following 2 variables.
  \describe{
    \item{\code{X1min}}{a numeric vector}
    \item{\code{X4weeks}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex09.44)
str(ex09.44)
}
\keyword{datasets}
