\name{ex09.65}
\alias{ex09.65}
\docType{data}
\title{R Data set: ex09.65}
\description{
The \code{ex09.65} data frame has 3 rows and 4 columns.
}
\usage{data(ex09.65)}
\format{
  A data frame with 3 observations on the following 4 variables.
  \describe{
    \item{\code{C1}}{a factor with levels \code{Fixed} \code{Floating} \code{Method}}
    \item{\code{C2}}{a factor with levels \code{10} \code{size}}
    \item{\code{C3}}{a factor with levels \code{757} \code{807} \code{mean}}
    \item{\code{C4}}{a factor with levels \code{27} \code{41} \code{SD}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex09.65)
str(ex09.65)
}
\keyword{datasets}
