\name{ex09.70}
\alias{ex09.70}
\docType{data}
\title{R Data set: ex09.70}
\description{
The \code{ex09.70} data frame has 5 rows and 4 columns.
}
\usage{data(ex09.70)}
\format{
  A data frame with 5 observations on the following 4 variables.
  \describe{
    \item{\code{C1}}{a factor with levels \code{} \code{C1} \code{Type} \code{Without side coating} \code{With side coating}}
    \item{\code{C2}}{a factor with levels \code{10} \code{C2} \code{Sample} \code{size}}
    \item{\code{C3}}{a factor with levels \code{63.23} \code{80.95} \code{C3} \code{mean} \code{Sample}}
    \item{\code{C4}}{a factor with levels \code{5.96} \code{9.59} \code{C4} \code{Sample} \code{SD}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex09.70)
str(ex09.70)
}
\keyword{datasets}
