\name{ex09.72}
\alias{ex09.72}
\docType{data}
\title{R Data set: ex09.72}
\description{
The \code{ex09.72} data frame has 17 rows and 3 columns.
}
\usage{data(ex09.72)}
\format{
  A data frame with 17 observations on the following 3 variables.
  \describe{
    \item{\code{Motor}}{a numeric vector}
    \item{\code{Commutator}}{a numeric vector}
    \item{\code{Pinion}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex09.72)
str(ex09.72)
}
\keyword{datasets}
