\name{ex09.76}
\alias{ex09.76}
\docType{data}
\title{R Data set: ex09.76}
\description{
The \code{ex09.76} data frame has 6 rows and 4 columns.
}
\usage{data(ex09.76)}
\format{
  A data frame with 6 observations on the following 4 variables.
  \describe{
    \item{\code{C1}}{a factor with levels \code{} \code{C1} \code{Clean} \code{Site} \code{Steam plant}}
    \item{\code{C2}}{a factor with levels \code{} \code{8} \code{9} \code{C2} \code{Sample} \code{size}}
    \item{\code{C3}}{a factor with levels \code{} \code{11} \code{18} \code{C3} \code{concentration} \code{Mean log}}
    \item{\code{C4}}{a factor with levels \code{4.6} \code{4.9} \code{C4} \code{concentration} \code{of log} \code{SD}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex09.76)
str(ex09.76)
}
\keyword{datasets}
