\name{ex09.77}
\alias{ex09.77}
\docType{data}
\title{R Data set: ex09.77}
\description{
The \code{ex09.77} data frame has 5 rows and 3 columns.
}
\usage{data(ex09.77)}
\format{
  A data frame with 5 observations on the following 3 variables.
  \describe{
    \item{\code{Twist.multiple}}{a numeric vector}
    \item{\code{Control.strength}}{a numeric vector}
    \item{\code{Heated.strength}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex09.77)
str(ex09.77)
}
\keyword{datasets}
