\name{ex09.78}
\alias{ex09.78}
\docType{data}
\title{R Data set: ex09.78}
\description{
The \code{ex09.78} data frame has 5 rows and 4 columns.
}
\usage{data(ex09.78)}
\format{
  A data frame with 5 observations on the following 4 variables.
  \describe{
    \item{\code{C1}}{a factor with levels \code{} \code{C1} \code{Elderly men} \code{Group} \code{Young}}
    \item{\code{C2}}{a factor with levels \code{12} \code{13} \code{C2} \code{Sample} \code{size}}
    \item{\code{C3}}{a factor with levels \code{6.71} \code{7.47} \code{C3} \code{mean} \code{Sample}}
    \item{\code{C4}}{a factor with levels \code{0.22} \code{0.28} \code{C4} \code{error} \code{Standard}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex09.78)
str(ex09.78)
}
\keyword{datasets}
