\name{ex09.79}
\alias{ex09.79}
\docType{data}
\title{R Data set: ex09.79}
\description{
The \code{ex09.79} data frame has 8 rows and 2 columns.
}
\usage{data(ex09.79)}
\format{
  A data frame with 8 observations on the following 2 variables.
  \describe{
    \item{\code{Good.visibility}}{a numeric vector}
    \item{\code{Poor.visibility}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex09.79)
str(ex09.79)
}
\keyword{datasets}
