\name{ex09.82}
\alias{ex09.82}
\docType{data}
\title{R Data set: ex09.82}
\description{
The \code{ex09.82} data frame has 7 rows and 2 columns.
}
\usage{data(ex09.82)}
\format{
  A data frame with 7 observations on the following 2 variables.
  \describe{
    \item{\code{expend}}{a numeric vector}
    \item{\code{intake}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex09.82)
str(ex09.82)
}
\keyword{datasets}
