\name{ex09.86}
\alias{ex09.86}
\docType{data}
\title{R Data set: ex09.86}
\description{
The \code{ex09.86} data frame has 4 rows and 3 columns.
}
\usage{data(ex09.86)}
\format{
  A data frame with 4 observations on the following 3 variables.
  \describe{
    \item{\code{Treatment}}{a numeric vector}
    \item{\code{n}}{a numeric vector}
    \item{\code{SD}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex09.86)
str(ex09.86)
}
\keyword{datasets}
