\name{ex09.90}
\alias{ex09.90}
\docType{data}
\title{R Data set: ex09.90}
\description{
The \code{ex09.90} data frame has 3 rows and 9 columns.
}
\usage{data(ex09.90)}
\format{
  A data frame with 3 observations on the following 9 variables.
  \describe{
    \item{\code{C1}}{a factor with levels \code{Frequency} \code{Region 1} \code{Region 2}}
    \item{\code{C2}}{a numeric vector}
    \item{\code{C3}}{a numeric vector}
    \item{\code{C4}}{a numeric vector}
    \item{\code{C5}}{a numeric vector}
    \item{\code{C6}}{a numeric vector}
    \item{\code{C7}}{a numeric vector}
    \item{\code{C8}}{a numeric vector}
    \item{\code{C9}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex09.90)
str(ex09.90)
}
\keyword{datasets}
