\name{ex09.92}
\alias{ex09.92}
\docType{data}
\title{R Data set: ex09.92}
\description{
The \code{ex09.92} data frame has 8 rows and 3 columns.
}
\usage{data(ex09.92)}
\format{
  A data frame with 8 observations on the following 3 variables.
  \describe{
    \item{\code{Number}}{a numeric vector}
    \item{\code{Region1}}{a numeric vector}
    \item{\code{Region2}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex09.92)
str(ex09.92)
}
\keyword{datasets}
