\name{ex10.06}
\alias{ex10.06}
\docType{data}
\title{R Data set: ex10.06}
\description{
The \code{ex10.06} data frame has 40 rows and 2 columns.
}
\usage{data(ex10.06)}
\format{
  A data frame with 40 observations on the following 2 variables.
  \describe{
    \item{\code{Fe}}{a numeric vector}
    \item{\code{formation.group}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex10.06)
str(ex10.06)
}
\keyword{datasets}
