\name{ex10.08}
\alias{ex10.08}
\docType{data}
\title{R Data set: ex10.08}
\description{
The \code{ex10.08} data frame has 35 rows and 2 columns.
}
\usage{data(ex10.08)}
\format{
  A data frame with 35 observations on the following 2 variables.
  \describe{
    \item{\code{stiffness}}{a numeric vector}
    \item{\code{plate.lengths}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex10.08)
str(ex10.08)
}
\keyword{datasets}
