\name{ex10.09}
\alias{ex10.09}
\docType{data}
\title{R Data set: ex10.09}
\description{
The \code{ex10.09} data frame has 24 rows and 2 columns.
}
\usage{data(ex10.09)}
\format{
  A data frame with 24 observations on the following 2 variables.
  \describe{
    \item{\code{thiamin}}{a numeric vector}
    \item{\code{type}}{a factor with levels \code{Barley} \code{Maize} \code{Oats} \code{Wheat}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex10.09)
str(ex10.09)
}
\keyword{datasets}
