\name{ex10.18}
\alias{ex10.18}
\docType{data}
\title{R Data set: ex10.18}
\description{
The \code{ex10.18} data frame has 4 rows and 5 columns.
}
\usage{data(ex10.18)}
\format{
  A data frame with 4 observations on the following 5 variables.
  \describe{
    \item{\code{Hormone.1}}{a numeric vector}
    \item{\code{Hormone.2}}{a numeric vector}
    \item{\code{Hormone.3}}{a numeric vector}
    \item{\code{Hormone.4}}{a numeric vector}
    \item{\code{Hormone.5}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex10.18)
str(ex10.18)
}
\keyword{datasets}
