\name{ex10.22}
\alias{ex10.22}
\docType{data}
\title{R Data set: ex10.22}
\description{
The \code{ex10.22} data frame has 5 rows and 4 columns.
}
\usage{data(ex10.22)}
\format{
  A data frame with 5 observations on the following 4 variables.
  \describe{
    \item{\code{level.1.6}}{a numeric vector}
    \item{\code{level.3.8}}{a numeric vector}
    \item{\code{level.6.0}}{a numeric vector}
    \item{\code{level.10.2}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex10.22)
str(ex10.22)
}
\keyword{datasets}
