\name{ex10.26}
\alias{ex10.26}
\docType{data}
\title{R Data set: ex10.26}
\description{
The \code{ex10.26} data frame has 5 rows and 6 columns.
}
\usage{data(ex10.26)}
\format{
  A data frame with 5 observations on the following 6 variables.
  \describe{
    \item{\code{Imperial}}{a numeric vector}
    \item{\code{Parkay}}{a numeric vector}
    \item{\code{Blue.Bonnet}}{a numeric vector}
    \item{\code{Chiffon}}{a numeric vector}
    \item{\code{Mazola}}{a numeric vector}
    \item{\code{Fleischmann}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex10.26)
str(ex10.26)
}
\keyword{datasets}
