\name{ex10.32}
\alias{ex10.32}
\docType{data}
\title{R Data set: ex10.32}
\description{
The \code{ex10.32} data frame has 5 rows and 4 columns.
}
\usage{data(ex10.32)}
\format{
  A data frame with 5 observations on the following 4 variables.
  \describe{
    \item{\code{A}}{a numeric vector}
    \item{\code{B}}{a numeric vector}
    \item{\code{C}}{a numeric vector}
    \item{\code{D}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex10.32)
str(ex10.32)
}
\keyword{datasets}
