\name{ex10.36}
\alias{ex10.36}
\docType{data}
\title{R Data set: ex10.36}
\description{
The \code{ex10.36} data frame has 4 rows and 5 columns.
}
\usage{data(ex10.36)}
\format{
  A data frame with 4 observations on the following 5 variables.
  \describe{
    \item{\code{L.D}}{a numeric vector}
    \item{\code{R}}{a numeric vector}
    \item{\code{R.L}}{a numeric vector}
    \item{\code{C}}{a numeric vector}
    \item{\code{C.L}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex10.36)
str(ex10.36)
}
\keyword{datasets}
