\name{ex10.37}
\alias{ex10.37}
\docType{data}
\title{R Data set: ex10.37}
\description{
The \code{ex10.37} data frame has 6 rows and 5 columns.
}
\usage{data(ex10.37)}
\format{
  A data frame with 6 observations on the following 5 variables.
  \describe{
    \item{\code{Brand.1}}{a numeric vector}
    \item{\code{Brand.2}}{a numeric vector}
    \item{\code{Brand.3}}{a numeric vector}
    \item{\code{Brand.4}}{a numeric vector}
    \item{\code{Brand.5}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex10.37)
str(ex10.37)
}
\keyword{datasets}
