\name{ex10.42}
\alias{ex10.42}
\docType{data}
\title{R Data set: ex10.42}
\description{
The \code{ex10.42} data frame has 19 rows and 2 columns.
}
\usage{data(ex10.42)}
\format{
  A data frame with 19 observations on the following 2 variables.
  \describe{
    \item{\code{cff}}{a numeric vector}
    \item{\code{color}}{a factor with levels \code{Blue} \code{Brown} \code{Green}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex10.42)
str(ex10.42)
}
\keyword{datasets}
