\name{ex11.03}
\alias{ex11.03}
\docType{data}
\title{R Data set: ex11.03}
\description{
The \code{ex11.03} data frame has 17 rows and 3 columns.
}
\usage{data(ex11.03)}
\format{
  A data frame with 17 observations on the following 3 variables.
  \describe{
    \item{\code{C1}}{a factor with levels \code{200} \code{226} \code{240} \code{261} \code{278} \code{312} \code{330} \code{369} \code{381} \code{416} \code{462} \code{500} \code{517} \code{575} \code{645} \code{733} \code{C1}}
    \item{\code{C2}}{a factor with levels \code{1(200)} \code{2(400)} \code{3(700)} \code{4(1100)} \code{C2}}
    \item{\code{C3}}{a factor with levels \code{1(190)} \code{2(250)} \code{3(300)} \code{4(400)} \code{C3}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.03)
str(ex11.03)
}
\keyword{datasets}
