\name{ex11.04}
\alias{ex11.04}
\docType{data}
\title{R Data set: ex11.04}
\description{
The \code{ex11.04} data frame has 12 rows and 3 columns.
}
\usage{data(ex11.04)}
\format{
  A data frame with 12 observations on the following 3 variables.
  \describe{
    \item{\code{Response}}{a numeric vector}
    \item{\code{Paint.Brand}}{a numeric vector}
    \item{\code{Roller.Brand}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.04)
str(ex11.04)
}
\keyword{datasets}
