\name{ex11.05}
\alias{ex11.05}
\docType{data}
\title{R Data set: ex11.05}
\description{
The \code{ex11.05} data frame has 20 rows and 3 columns.
}
\usage{data(ex11.05)}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{force}}{a numeric vector}
    \item{\code{connector}}{a numeric vector}
    \item{\code{angle}}{a factor with levels \code{0 deg} \code{2 deg} \code{4 deg} \code{6 deg}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.05)
str(ex11.05)
}
\keyword{datasets}
