\name{ex11.09}
\alias{ex11.09}
\docType{data}
\title{R Data set: ex11.09}
\description{
The \code{ex11.09} data frame has 36 rows and 3 columns.
}
\usage{data(ex11.09)}
\format{
  A data frame with 36 observations on the following 3 variables.
  \describe{
    \item{\code{response}}{a numeric vector}
    \item{\code{type}}{a numeric vector}
    \item{\code{subject}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.09)
str(ex11.09)
}
\keyword{datasets}
