\name{ex11.10}
\alias{ex11.10}
\docType{data}
\title{R Data set: ex11.10}
\description{
The \code{ex11.10} data frame has 10 rows and 4 columns.
}
\usage{data(ex11.10)}
\format{
  A data frame with 10 observations on the following 4 variables.
  \describe{
    \item{\code{Batch}}{a numeric vector}
    \item{\code{Method.A}}{a numeric vector}
    \item{\code{Method.B}}{a numeric vector}
    \item{\code{Method.C}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.10)
str(ex11.10)
}
\keyword{datasets}
