\name{ex11.15}
\alias{ex11.15}
\docType{data}
\title{R Data set: ex11.15}
\description{
The \code{ex11.15} data frame has 18 rows and 4 columns.
}
\usage{data(ex11.15)}
\format{
  A data frame with 18 observations on the following 4 variables.
  \describe{
    \item{\code{Sand}}{a factor with levels \code{0} \code{15} \code{30}}
    \item{\code{Carbon}}{a factor with levels \code{0} \code{0.25} \code{0.5}}
    \item{\code{Hardness}}{a numeric vector}
    \item{\code{Strength}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.15)
str(ex11.15)
}
\keyword{datasets}
