\name{ex11.16}
\alias{ex11.16}
\docType{data}
\title{R Data set: ex11.16}
\description{
The \code{ex11.16} data frame has 18 rows and 3 columns.
}
\usage{data(ex11.16)}
\format{
  A data frame with 18 observations on the following 3 variables.
  \describe{
    \item{\code{Response}}{a numeric vector}
    \item{\code{Formulat}}{a factor with levels \code{1} \code{2}}
    \item{\code{Speed}}{a factor with levels \code{60} \code{70} \code{80}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.16)
str(ex11.16)
}
\keyword{datasets}
