\name{ex11.17}
\alias{ex11.17}
\docType{data}
\title{R Data set: ex11.17}
\description{
The \code{ex11.17} data frame has 18 rows and 4 columns.
}
\usage{data(ex11.17)}
\format{
  A data frame with 18 observations on the following 4 variables.
  \describe{
    \item{\code{Sand.Addition.perc}}{a numeric vector}
    \item{\code{Carbon.Fiber.Addition.perc}}{a numeric vector}
    \item{\code{Casting.hardness}}{a numeric vector}
    \item{\code{Wet.Mold.Strength}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.17)
str(ex11.17)
}
\keyword{datasets}
