\name{ex11.18}
\alias{ex11.18}
\docType{data}
\title{R Data set: ex11.18}
\description{
The \code{ex11.18} data frame has 18 rows and 3 columns.
}
\usage{data(ex11.18)}
\format{
  A data frame with 18 observations on the following 3 variables.
  \describe{
    \item{\code{Yield}}{a numeric vector}
    \item{\code{Speed}}{a numeric vector}
    \item{\code{Formulation}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.18)
str(ex11.18)
}
\keyword{datasets}
