\name{ex11.29}
\alias{ex11.29}
\docType{data}
\title{R Data set: ex11.29}
\description{
The \code{ex11.29} data frame has 96 rows and 4 columns.
}
\usage{data(ex11.29)}
\format{
  A data frame with 96 observations on the following 4 variables.
  \describe{
    \item{\code{length}}{a numeric vector}
    \item{\code{time}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{heat}}{a factor with levels \code{1} \code{2}}
    \item{\code{machine}}{a factor with levels \code{1} \code{2} \code{3} \code{4}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.29)
str(ex11.29)
}
\keyword{datasets}
