\name{ex11.31}
\alias{ex11.31}
\docType{data}
\title{R Data set: ex11.31}
\description{
The \code{ex11.31} data frame has 27 rows and 4 columns.
}
\usage{data(ex11.31)}
\format{
  A data frame with 27 observations on the following 4 variables.
  \describe{
    \item{\code{Yield}}{a numeric vector}
    \item{\code{time}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{tempture}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{pressure}}{a factor with levels \code{1} \code{2} \code{3}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.31)
str(ex11.31)
}
\keyword{datasets}
