\name{ex11.34}
\alias{ex11.34}
\docType{data}
\title{R Data set: ex11.34}
\description{
The \code{ex11.34} data frame has 36 rows and 4 columns.
}
\usage{data(ex11.34)}
\format{
  A data frame with 36 observations on the following 4 variables.
  \describe{
    \item{\code{Sales}}{a numeric vector}
    \item{\code{store}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6}}
    \item{\code{week}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6}}
    \item{\code{shelf}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.34)
str(ex11.34)
}
\keyword{datasets}
