\name{ex11.35}
\alias{ex11.35}
\docType{data}
\title{R Data set: ex11.35}
\description{
The \code{ex11.35} data frame has 25 rows and 4 columns.
}
\usage{data(ex11.35)}
\format{
  A data frame with 25 observations on the following 4 variables.
  \describe{
    \item{\code{Moisture}}{a numeric vector}
    \item{\code{plant}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5}}
    \item{\code{leafsize}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5}}
    \item{\code{time}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.35)
str(ex11.35)
}
\keyword{datasets}
