\name{ex11.39}
\alias{ex11.39}
\docType{data}
\title{R Data set: ex11.39}
\description{
The \code{ex11.39} data frame has 24 rows and 4 columns.
}
\usage{data(ex11.39)}
\format{
  A data frame with 24 observations on the following 4 variables.
  \describe{
    \item{\code{cleaning}}{a numeric vector}
    \item{\code{detergnt}}{a factor with levels \code{1} \code{2}}
    \item{\code{carbonat}}{a factor with levels \code{1} \code{2}}
    \item{\code{cellulos}}{a factor with levels \code{1} \code{2}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.39)
str(ex11.39)
}
\keyword{datasets}
