\name{ex11.40}
\alias{ex11.40}
\docType{data}
\title{R Data set: ex11.40}
\description{
The \code{ex11.40} data frame has 32 rows and 5 columns.
}
\usage{data(ex11.40)}
\format{
  A data frame with 32 observations on the following 5 variables.
  \describe{
    \item{\code{sizing}}{a numeric vector}
    \item{\code{conc}}{a factor with levels \code{50} \code{75}}
    \item{\code{pH}}{a factor with levels \code{6} \code{7}}
    \item{\code{tempture}}{a factor with levels \code{60} \code{70}}
    \item{\code{time}}{a factor with levels \code{6} \code{8}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.40)
str(ex11.40)
}
\keyword{datasets}
