\name{ex11.42}
\alias{ex11.42}
\docType{data}
\title{R Data set: ex11.42}
\description{
The \code{ex11.42} data frame has 48 rows and 5 columns.
}
\usage{data(ex11.42)}
\format{
  A data frame with 48 observations on the following 5 variables.
  \describe{
    \item{\code{consump}}{a numeric vector}
    \item{\code{roof}}{a factor with levels \code{-1} \code{1}}
    \item{\code{power}}{a factor with levels \code{-1} \code{1}}
    \item{\code{scrap}}{a factor with levels \code{-1} \code{1}}
    \item{\code{charge}}{a factor with levels \code{-1} \code{1}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.42)
str(ex11.42)
}
\keyword{datasets}
