\name{ex11.43}
\alias{ex11.43}
\docType{data}
\title{R Data set: ex11.43}
\description{
The \code{ex11.43} data frame has 16 rows and 5 columns.
}
\usage{data(ex11.43)}
\format{
  A data frame with 16 observations on the following 5 variables.
  \describe{
    \item{\code{duration}}{a numeric vector}
    \item{\code{vibratn}}{a factor with levels \code{-1} \code{1}}
    \item{\code{tempture}}{a factor with levels \code{-1} \code{1}}
    \item{\code{altitude}}{a factor with levels \code{-1} \code{1}}
    \item{\code{firing}}{a factor with levels \code{-1} \code{1}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.43)
str(ex11.43)
}
\keyword{datasets}
