\name{ex11.50}
\alias{ex11.50}
\docType{data}
\title{R Data set: ex11.50}
\description{
The \code{ex11.50} data frame has 45 rows and 3 columns.
}
\usage{data(ex11.50)}
\format{
  A data frame with 45 observations on the following 3 variables.
  \describe{
    \item{\code{Fabric}}{a factor with levels \code{Broadcloth} \code{Corduroy} \code{Crepe} \code{Denim} \code{Double knit} \code{Sheeting} \code{Terry} \code{Twill} \code{Twill mix}}
    \item{\code{Response}}{a numeric vector}
    \item{\code{Drying}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.50)
str(ex11.50)
}
\keyword{datasets}
