\name{ex11.52}
\alias{ex11.52}
\docType{data}
\title{R Data set: ex11.52}
\description{
The \code{ex11.52} data frame has 16 rows and 3 columns.
}
\usage{data(ex11.52)}
\format{
  A data frame with 16 observations on the following 3 variables.
  \describe{
    \item{\code{Response}}{a numeric vector}
    \item{\code{Sowing.Rate.kg.ha}}{a numeric vector}
    \item{\code{Plot}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.52)
str(ex11.52)
}
\keyword{datasets}
