\name{ex11.53}
\alias{ex11.53}
\docType{data}
\title{R Data set: ex11.53}
\description{
The \code{ex11.53} data frame has 8 rows and 6 columns.
}
\usage{data(ex11.53)}
\format{
  A data frame with 8 observations on the following 6 variables.
  \describe{
    \item{\code{Run}}{a numeric vector}
    \item{\code{Spray.Volume}}{a factor with levels \code{-} \code{+}}
    \item{\code{Belt.Speed}}{a factor with levels \code{-} \code{+}}
    \item{\code{Brand}}{a factor with levels \code{-} \code{+}}
    \item{\code{Replication.1}}{a numeric vector}
    \item{\code{Replication.2}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.53)
str(ex11.53)
}
\keyword{datasets}
