\name{ex11.54}
\alias{ex11.54}
\docType{data}
\title{R Data set: ex11.54}
\description{
The \code{ex11.54} data frame has 8 rows and 5 columns.
}
\usage{data(ex11.54)}
\format{
  A data frame with 8 observations on the following 5 variables.
  \describe{
    \item{\code{Sample.number}}{a numeric vector}
    \item{\code{Factor.A}}{a numeric vector}
    \item{\code{Factor.B}}{a numeric vector}
    \item{\code{Factor.C}}{a numeric vector}
    \item{\code{Resonse.EC50}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.54)
str(ex11.54)
}
\keyword{datasets}
