\name{ex11.55}
\alias{ex11.55}
\docType{data}
\title{R Data set: ex11.55}
\description{
The \code{ex11.55} data frame has 16 rows and 2 columns.
}
\usage{data(ex11.55)}
\format{
  A data frame with 16 observations on the following 2 variables.
  \describe{
    \item{\code{Test.Run}}{a numeric vector}
    \item{\code{Iron.Extraction}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.55)
str(ex11.55)
}
\keyword{datasets}
