\name{ex11.56}
\alias{ex11.56}
\docType{data}
\title{R Data set: ex11.56}
\description{
The \code{ex11.56} data frame has 30 rows and 3 columns.
}
\usage{data(ex11.56)}
\format{
  A data frame with 30 observations on the following 3 variables.
  \describe{
    \item{\code{C1}}{a numeric vector}
    \item{\code{C2}}{a factor with levels \code{pH 3} \code{pH 5.5} \code{pH 7}}
    \item{\code{C3}}{a factor with levels \code{Diseased} \code{Healthy}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.56)
str(ex11.56)
}
\keyword{datasets}
