\name{ex11.57}
\alias{ex11.57}
\docType{data}
\title{R Data set: ex11.57}
\description{
The \code{ex11.57} data frame has 54 rows and 4 columns.
}
\usage{data(ex11.57)}
\format{
  A data frame with 54 observations on the following 4 variables.
  \describe{
    \item{\code{Response}}{a numeric vector}
    \item{\code{Pressure}}{a factor with levels \code{Pressure 103.4} \code{Pressure17.2} \code{Pressure34.4}}
    \item{\code{Temp}}{a factor with levels \code{50 degrees} \code{75 degrees} \code{8 degrees}}
    \item{\code{Fabric}}{a factor with levels \code{420-D} \code{630-D} \code{840-D}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.57)
str(ex11.57)
}
\keyword{datasets}
