\name{ex11.59}
\alias{ex11.59}
\docType{data}
\title{R Data set: ex11.59}
\description{
The \code{ex11.59} data frame has 36 rows and 4 columns.
}
\usage{data(ex11.59)}
\format{
  A data frame with 36 observations on the following 4 variables.
  \describe{
    \item{\code{Cure.Time.1}}{a numeric vector}
    \item{\code{Adhesive.type}}{a factor with levels \code{Copper} \code{Nickel}}
    \item{\code{Adhesive.factor}}{a numeric vector}
    \item{\code{Cure.Time}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.59)
str(ex11.59)
}
\keyword{datasets}
