\name{ex11.61}
\alias{ex11.61}
\docType{data}
\title{R Data set: ex11.61}
\description{
The \code{ex11.61} data frame has 25 rows and 5 columns.
}
\usage{data(ex11.61)}
\format{
  A data frame with 25 observations on the following 5 variables.
  \describe{
    \item{\code{weight}}{a numeric vector}
    \item{\code{volume}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5}}
    \item{\code{color}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5}}
    \item{\code{size}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5}}
    \item{\code{time}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex11.61)
str(ex11.61)
}
\keyword{datasets}
