\name{ex12.02}
\alias{ex12.02}
\docType{data}
\title{R Data set: ex12.02}
\description{
The \code{ex12.02} data frame has 10 rows and 4 columns.
}
\usage{data(ex12.02)}
\format{
  A data frame with 10 observations on the following 4 variables.
  \describe{
    \item{\code{Engine}}{a numeric vector}
    \item{\code{Age}}{a numeric vector}
    \item{\code{Baseline}}{a numeric vector}
    \item{\code{Reformulated}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex12.02)
str(ex12.02)
}
\keyword{datasets}
