\name{ex12.24}
\alias{ex12.24}
\docType{data}
\title{R Data set: ex12.24}
\description{
The \code{ex12.24} data frame has 6 rows and 2 columns.
}
\usage{data(ex12.24)}
\format{
  A data frame with 6 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex12.24)
str(ex12.24)
}
\keyword{datasets}
