\name{ex12.71}
\alias{ex12.71}
\docType{data}
\title{R Data set: ex12.71}
\description{
The \code{ex12.71} data frame has 17 rows and 2 columns.
}
\usage{data(ex12.71)}
\format{
  A data frame with 17 observations on the following 2 variables.
  \describe{
    \item{\code{X}}{a numeric vector}
    \item{\code{Y}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex12.71)
str(ex12.71)
}
\keyword{datasets}
