\name{ex12.72}
\alias{ex12.72}
\docType{data}
\title{R Data set: ex12.72}
\description{
The \code{ex12.72} data frame has 9 rows and 2 columns.
}
\usage{data(ex12.72)}
\format{
  A data frame with 9 observations on the following 2 variables.
  \describe{
    \item{\code{CO}}{a numeric vector}
    \item{\code{NO3}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex12.72)
str(ex12.72)
}
\keyword{datasets}
