\name{ex13.09}
\alias{ex13.09}
\docType{data}
\title{R Data set: ex13.09}
\description{
The \code{ex13.09} data frame has 44 rows and 3 columns.
}
\usage{data(ex13.09)}
\format{
  A data frame with 44 observations on the following 3 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
    \item{\code{set}}{a factor with levels \code{a} \code{b} \code{c} \code{d}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex13.09)
str(ex13.09)
}
\keyword{datasets}
