\name{ex13.09b}
\alias{ex13.09b}
\docType{data}
\title{R Data set: ex13.09b}
\description{
The \code{ex13.09b} data frame has 11 rows and 2 columns.
}
\usage{data(ex13.09b)}
\format{
  A data frame with 11 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex13.09b)
str(ex13.09b)
}
\keyword{datasets}
