\name{ex13.14}
\alias{ex13.14}
\docType{data}
\title{R Data set: ex13.14}
\description{
The \code{ex13.14} data frame has 14 rows and 2 columns.
}
\usage{data(ex13.14)}
\format{
  A data frame with 14 observations on the following 2 variables.
  \describe{
    \item{\code{Col1}}{a numeric vector}
    \item{\code{Col2}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex13.14)
str(ex13.14)
}
\keyword{datasets}
