\name{ex13.24}
\alias{ex13.24}
\docType{data}
\title{R Data set: ex13.24}
\description{
The \code{ex13.24} data frame has 22 rows and 2 columns.
}
\usage{data(ex13.24)}
\format{
  A data frame with 22 observations on the following 2 variables.
  \describe{
    \item{\code{Kyphosis}}{a numeric vector}
    \item{\code{No.kyphosis}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex13.24)
str(ex13.24)
}
\keyword{datasets}
