\name{ex13.31}
\alias{ex13.31}
\docType{data}
\title{R Data set: ex13.31}
\description{
The \code{ex13.31} data frame has 7 rows and 2 columns.
}
\usage{data(ex13.31)}
\format{
  A data frame with 7 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex13.31)
str(ex13.31)
}
\keyword{datasets}
