\name{ex13.47}
\alias{ex13.47}
\docType{data}
\title{R Data set: ex13.47}
\description{
The \code{ex13.47} data frame has 30 rows and 6 columns.
}
\usage{data(ex13.47)}
\format{
  A data frame with 30 observations on the following 6 variables.
  \describe{
    \item{\code{Row}}{a numeric vector}
    \item{\code{Plastics}}{a numeric vector}
    \item{\code{Paper}}{a numeric vector}
    \item{\code{Garbage}}{a numeric vector}
    \item{\code{Water}}{a numeric vector}
    \item{\code{Energy.content}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex13.47)
str(ex13.47)
}
\keyword{datasets}
