\name{ex13.48}
\alias{ex13.48}
\docType{data}
\title{R Data set: ex13.48}
\description{
The \code{ex13.48} data frame has 15 rows and 4 columns.
}
\usage{data(ex13.48)}
\format{
  A data frame with 15 observations on the following 4 variables.
  \describe{
    \item{\code{x1}}{a numeric vector}
    \item{\code{x2}}{a numeric vector}
    \item{\code{x3}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex13.48)
str(ex13.48)
}
\keyword{datasets}
