\name{ex13.50}
\alias{ex13.50}
\docType{data}
\title{R Data set: ex13.50}
\description{
The \code{ex13.50} data frame has 14 rows and 3 columns.
}
\usage{data(ex13.50)}
\format{
  A data frame with 14 observations on the following 3 variables.
  \describe{
    \item{\code{y}}{a numeric vector}
    \item{\code{x1}}{a numeric vector}
    \item{\code{x2}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex13.50)
str(ex13.50)
}
\keyword{datasets}
