\name{ex13.52}
\alias{ex13.52}
\docType{data}
\title{R Data set: ex13.52}
\description{
The \code{ex13.52} data frame has 20 rows and 4 columns.
}
\usage{data(ex13.52)}
\format{
  A data frame with 20 observations on the following 4 variables.
  \describe{
    \item{\code{Linoleic}}{a numeric vector}
    \item{\code{Kerosene}}{a numeric vector}
    \item{\code{Antiox}}{a numeric vector}
    \item{\code{Betacaro}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex13.52)
str(ex13.52)
}
\keyword{datasets}
