\name{ex13.54}
\alias{ex13.54}
\docType{data}
\title{R Data set: ex13.54}
\description{
The \code{ex13.54} data frame has 31 rows and 5 columns.
}
\usage{data(ex13.54)}
\format{
  A data frame with 31 observations on the following 5 variables.
  \describe{
    \item{\code{Bright}}{a numeric vector}
    \item{\code{H2O2}}{a numeric vector}
    \item{\code{NaOH}}{a numeric vector}
    \item{\code{Silicate}}{a numeric vector}
    \item{\code{Tempture}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex13.54)
str(ex13.54)
}
\keyword{datasets}
