\name{ex13.55}
\alias{ex13.55}
\docType{data}
\title{R Data set: ex13.55}
\description{
The \code{ex13.55} data frame has 10 rows and 3 columns.
}
\usage{data(ex13.55)}
\format{
  A data frame with 10 observations on the following 3 variables.
  \describe{
    \item{\code{q}}{a numeric vector}
    \item{\code{a}}{a numeric vector}
    \item{\code{b}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex13.55)
str(ex13.55)
}
\keyword{datasets}
