\name{ex13.66}
\alias{ex13.66}
\docType{data}
\title{R Data set: ex13.66}
\description{
The \code{ex13.66} data frame has 9 rows and 3 columns.
}
\usage{data(ex13.66)}
\format{
  A data frame with 9 observations on the following 3 variables.
  \describe{
    \item{\code{x1.in}}{a numeric vector}
    \item{\code{x2.in}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex13.66)
str(ex13.66)
}
\keyword{datasets}
