\name{ex13.67}
\alias{ex13.67}
\docType{data}
\title{R Data set: ex13.67}
\description{
The \code{ex13.67} data frame has 32 rows and 7 columns.
}
\usage{data(ex13.67)}
\format{
  A data frame with 32 observations on the following 7 variables.
  \describe{
    \item{\code{Obs}}{a numeric vector}
    \item{\code{pdconc}}{a numeric vector}
    \item{\code{niconc}}{a numeric vector}
    \item{\code{pH}}{a numeric vector}
    \item{\code{temp}}{a numeric vector}
    \item{\code{currdens}}{a numeric vector}
    \item{\code{pallcont}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex13.67)
str(ex13.67)
}
\keyword{datasets}
