\name{ex13.74}
\alias{ex13.74}
\docType{data}
\title{R Data set: ex13.74}
\description{
The \code{ex13.74} data frame has 12 rows and 2 columns.
}
\usage{data(ex13.74)}
\format{
  A data frame with 12 observations on the following 2 variables.
  \describe{
    \item{\code{log.con}}{a numeric vector}
    \item{\code{Li2O}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex13.74)
str(ex13.74)
}
\keyword{datasets}
