\name{ex13.75}
\alias{ex13.75}
\docType{data}
\title{R Data set: ex13.75}
\description{
The \code{ex13.75} data frame has 10 rows and 2 columns.
}
\usage{data(ex13.75)}
\format{
  A data frame with 10 observations on the following 2 variables.
  \describe{
    \item{\code{height}}{a numeric vector}
    \item{\code{log.Mn}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex13.75)
str(ex13.75)
}
\keyword{datasets}
