\name{ex14.13}
\alias{ex14.13}
\docType{data}
\title{R Data set: ex14.13}
\description{
The \code{ex14.13} data frame has 3 rows and 2 columns.
}
\usage{data(ex14.13)}
\format{
  A data frame with 3 observations on the following 2 variables.
  \describe{
    \item{\code{ovaries.developed}}{a numeric vector}
    \item{\code{Observed.count}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex14.13)
str(ex14.13)
}
\keyword{datasets}
