\name{ex14.18}
\alias{ex14.18}
\docType{data}
\title{R Data set: ex14.18}
\description{
The \code{ex14.18} data frame has 5 rows and 2 columns.
}
\usage{data(ex14.18)}
\format{
  A data frame with 5 observations on the following 2 variables.
  \describe{
    \item{\code{Rate.per.day}}{a factor with levels \code{<100} \code{.100-below .150} \code{.150-below .200} \code{.200-below .250} \code{.250 or more}}
    \item{\code{Frequency}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex14.18)
str(ex14.18)
}
\keyword{datasets}
