\name{ex14.26}
\alias{ex14.26}
\docType{data}
\title{R Data set: ex14.26}
\description{
The \code{ex14.26} data frame has 7 rows and 3 columns.
}
\usage{data(ex14.26)}
\format{
  A data frame with 7 observations on the following 3 variables.
  \describe{
    \item{\code{C1}}{a factor with levels \code{C1} \code{Control} \code{Eight leaves removed} \code{Four leaves removed} \code{Six leaves removed} \code{Treatment} \code{Two leaves removed}}
    \item{\code{C2}}{a factor with levels \code{141} \code{20} \code{24} \code{25} \code{28} \code{C2} \code{Matured}}
    \item{\code{C3}}{a factor with levels \code{206} \code{69} \code{73} \code{78} \code{82} \code{Aborted} \code{C3}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex14.26)
str(ex14.26)
}
\keyword{datasets}
