\name{ex14.27}
\alias{ex14.27}
\docType{data}
\title{R Data set: ex14.27}
\description{
The \code{ex14.27} data frame has 2 rows and 5 columns.
}
\usage{data(ex14.27)}
\format{
  A data frame with 2 observations on the following 5 variables.
  \describe{
    \item{\code{C1}}{a factor with levels \code{Men} \code{Women}}
    \item{\code{L.R}}{a numeric vector}
    \item{\code{L.R.1}}{a numeric vector}
    \item{\code{L.R.2}}{a numeric vector}
    \item{\code{Sample.size}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex14.27)
str(ex14.27)
}
\keyword{datasets}
