\name{ex14.28}
\alias{ex14.28}
\docType{data}
\title{R Data set: ex14.28}
\description{
The \code{ex14.28} data frame has 4 rows and 5 columns.
}
\usage{data(ex14.28)}
\format{
  A data frame with 4 observations on the following 5 variables.
  \describe{
    \item{\code{Trreatment}}{a factor with levels \code{Sham} \code{Solvent} \code{Thienylalanine} \code{Unhandled}}
    \item{\code{No.response}}{a numeric vector}
    \item{\code{Wild.running}}{a numeric vector}
    \item{\code{Clonic.seizure}}{a numeric vector}
    \item{\code{Tonic.seizure}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex14.28)
str(ex14.28)
}
\keyword{datasets}
