\name{ex14.29}
\alias{ex14.29}
\docType{data}
\title{R Data set: ex14.29}
\description{
The \code{ex14.29} data frame has 6 rows and 4 columns.
}
\usage{data(ex14.29)}
\format{
  A data frame with 6 observations on the following 4 variables.
  \describe{
    \item{\code{Male.genotype}}{a numeric vector}
    \item{\code{M.M}}{a numeric vector}
    \item{\code{M.F}}{a numeric vector}
    \item{\code{F.F}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex14.29)
str(ex14.29)
}
\keyword{datasets}
