\name{ex14.30}
\alias{ex14.30}
\docType{data}
\title{R Data set: ex14.30}
\description{
The \code{ex14.30} data frame has 4 rows and 5 columns.
}
\usage{data(ex14.30)}
\format{
  A data frame with 4 observations on the following 5 variables.
  \describe{
    \item{\code{C1}}{a factor with levels \code{1} \code{2} \code{3} \code{Configuration}}
    \item{\code{C2}}{a numeric vector}
    \item{\code{C3}}{a numeric vector}
    \item{\code{C4}}{a numeric vector}
    \item{\code{C5}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex14.30)
str(ex14.30)
}
\keyword{datasets}
