\name{ex14.31}
\alias{ex14.31}
\docType{data}
\title{R Data set: ex14.31}
\description{
The \code{ex14.31} data frame has 12 rows and 3 columns.
}
\usage{data(ex14.31)}
\format{
  A data frame with 12 observations on the following 3 variables.
  \describe{
    \item{\code{count}}{a numeric vector}
    \item{\code{Size}}{a factor with levels \code{Compact} \code{Fullsize} \code{Midsize} \code{Subcompact}}
    \item{\code{dist}}{a factor with levels \code{0-<10} \code{10-<20} \code{>=20}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex14.31)
str(ex14.31)
}
\keyword{datasets}
