\name{ex14.38}
\alias{ex14.38}
\docType{data}
\title{R Data set: ex14.38}
\description{
The \code{ex14.38} data frame has 3 rows and 3 columns.
}
\usage{data(ex14.38)}
\format{
  A data frame with 3 observations on the following 3 variables.
  \describe{
    \item{\code{Treatment}}{a factor with levels \code{Control} \code{New oil} \code{Old oil}}
    \item{\code{Parasitized}}{a numeric vector}
    \item{\code{Nonparasitized}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex14.38)
str(ex14.38)
}
\keyword{datasets}
