\name{ex14.40}
\alias{ex14.40}
\docType{data}
\title{R Data set: ex14.40}
\description{
The \code{ex14.40} data frame has 6 rows and 3 columns.
}
\usage{data(ex14.40)}
\format{
  A data frame with 6 observations on the following 3 variables.
  \describe{
    \item{\code{C1}}{a factor with levels \code{Baseball} \code{Basketball} \code{C1} \code{Football} \code{Hockey} \code{Sport}}
    \item{\code{C2}}{a factor with levels \code{150} \code{65} \code{72} \code{86} \code{C2} \code{Leader Wins}}
    \item{\code{C3}}{a factor with levels \code{15} \code{21} \code{39} \code{6} \code{C3} \code{Leader Loses}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex14.40)
str(ex14.40)
}
\keyword{datasets}
