\name{ex14.42}
\alias{ex14.42}
\docType{data}
\title{R Data set: ex14.42}
\description{
The \code{ex14.42} data frame has 4 rows and 4 columns.
}
\usage{data(ex14.42)}
\format{
  A data frame with 4 observations on the following 4 variables.
  \describe{
    \item{\code{Age}}{a factor with levels \code{15-54} \code{55-64} \code{65-74} \code{Over 74}}
    \item{\code{Home}}{a numeric vector}
    \item{\code{Acute}}{a numeric vector}
    \item{\code{Chronic}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex14.42)
str(ex14.42)
}
\keyword{datasets}
