\name{ex14.44}
\alias{ex14.44}
\docType{data}
\title{R Data set: ex14.44}
\description{
The \code{ex14.44} data frame has 4 rows and 6 columns.
}
\usage{data(ex14.44)}
\format{
  A data frame with 4 observations on the following 6 variables.
  \describe{
    \item{\code{C1}}{a factor with levels \code{Age} \code{C1} \code{Number in Sample} \code{Number who want item pricing}}
    \item{\code{C2}}{a factor with levels \code{127} \code{150} \code{<30} \code{C2}}
    \item{\code{C3}}{a factor with levels \code{118} \code{141} \code{30-39} \code{C3}}
    \item{\code{C4}}{a factor with levels \code{40-49} \code{77} \code{82} \code{C4}}
    \item{\code{C5}}{a factor with levels \code{50-59} \code{61} \code{63} \code{C5}}
    \item{\code{C6}}{a factor with levels \code{41} \code{49} \code{>60} \code{C6}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex14.44)
str(ex14.44)
}
\keyword{datasets}
