\name{ex15.03}
\alias{ex15.03}
\docType{data}
\title{R Data set: ex15.03}
\description{
The \code{ex15.03} data frame has 14 rows and 1 column.
}
\usage{data(ex15.03)}
\format{
  A data frame with 14 observations on the following variable.
  \describe{
    \item{\code{C1}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex15.03)
str(ex15.03)
}
\keyword{datasets}
