\name{ex15.05}
\alias{ex15.05}
\docType{data}
\title{R Data set: ex15.05}
\description{
The \code{ex15.05} data frame has 12 rows and 3 columns.
}
\usage{data(ex15.05)}
\format{
  A data frame with 12 observations on the following 3 variables.
  \describe{
    \item{\code{Sample}}{a numeric vector}
    \item{\code{Gravimetric}}{a numeric vector}
    \item{\code{Spectrophotometric}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex15.05)
str(ex15.05)
}
\keyword{datasets}
