\name{ex15.12}
\alias{ex15.12}
\docType{data}
\title{R Data set: ex15.12}
\description{
The \code{ex15.12} data frame has 8 rows and 2 columns.
}
\usage{data(ex15.12)}
\format{
  A data frame with 8 observations on the following 2 variables.
  \describe{
    \item{\code{Original.process}}{a numeric vector}
    \item{\code{Modified.process}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex15.12)
str(ex15.12)
}
\keyword{datasets}
