\name{ex15.14}
\alias{ex15.14}
\docType{data}
\title{R Data set: ex15.14}
\description{
The \code{ex15.14} data frame has 10 rows and 2 columns.
}
\usage{data(ex15.14)}
\format{
  A data frame with 10 observations on the following 2 variables.
  \describe{
    \item{\code{Orange.juice}}{a numeric vector}
    \item{\code{Ascorbic.acid}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex15.14)
str(ex15.14)
}
\keyword{datasets}
