\name{ex15.23}
\alias{ex15.23}
\docType{data}
\title{R Data set: ex15.23}
\description{
The \code{ex15.23} data frame has 5 rows and 4 columns.
}
\usage{data(ex15.23)}
\format{
  A data frame with 5 observations on the following 4 variables.
  \describe{
    \item{\code{Region.1}}{a numeric vector}
    \item{\code{Region.2}}{a numeric vector}
    \item{\code{Region.3}}{a numeric vector}
    \item{\code{Region.4}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex15.23)
str(ex15.23)
}
\keyword{datasets}
