\name{ex15.24}
\alias{ex15.24}
\docType{data}
\title{R Data set: ex15.24}
\description{
The \code{ex15.24} data frame has 9 rows and 4 columns.
}
\usage{data(ex15.24)}
\format{
  A data frame with 9 observations on the following 4 variables.
  \describe{
    \item{\code{fasting}}{a numeric vector}
    \item{\code{X23.protein}}{a numeric vector}
    \item{\code{X32.protein}}{a numeric vector}
    \item{\code{X67.protein}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex15.24)
str(ex15.24)
}
\keyword{datasets}
