\name{ex15.26}
\alias{ex15.26}
\docType{data}
\title{R Data set: ex15.26}
\description{
The \code{ex15.26} data frame has 10 rows and 5 columns.
}
\usage{data(ex15.26)}
\format{
  A data frame with 10 observations on the following 5 variables.
  \describe{
    \item{\code{Blocks}}{a numeric vector}
    \item{\code{A}}{a numeric vector}
    \item{\code{B}}{a numeric vector}
    \item{\code{C}}{a numeric vector}
    \item{\code{D}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex15.26)
str(ex15.26)
}
\keyword{datasets}
