\name{ex15.27}
\alias{ex15.27}
\docType{data}
\title{R Data set: ex15.27}
\description{
The \code{ex15.27} data frame has 10 rows and 4 columns.
}
\usage{data(ex15.27)}
\format{
  A data frame with 10 observations on the following 4 variables.
  \describe{
    \item{\code{Dog}}{a numeric vector}
    \item{\code{Isoflurane}}{a numeric vector}
    \item{\code{Halothane}}{a numeric vector}
    \item{\code{Cyclopropane}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex15.27)
str(ex15.27)
}
\keyword{datasets}
