\name{ex15.28}
\alias{ex15.28}
\docType{data}
\title{R Data set: ex15.28}
\description{
The \code{ex15.28} data frame has 8 rows and 3 columns.
}
\usage{data(ex15.28)}
\format{
  A data frame with 8 observations on the following 3 variables.
  \describe{
    \item{\code{Subject}}{a numeric vector}
    \item{\code{Potato}}{a numeric vector}
    \item{\code{Rice}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex15.28)
str(ex15.28)
}
\keyword{datasets}
