\name{ex15.30}
\alias{ex15.30}
\docType{data}
\title{R Data set: ex15.30}
\description{
The \code{ex15.30} data frame has 5 rows and 4 columns.
}
\usage{data(ex15.30)}
\format{
  A data frame with 5 observations on the following 4 variables.
  \describe{
    \item{\code{Treatment.I}}{a numeric vector}
    \item{\code{Treatment.II}}{a numeric vector}
    \item{\code{Treatment.III}}{a numeric vector}
    \item{\code{Treatment.IV}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex15.30)
str(ex15.30)
}
\keyword{datasets}
