\name{ex15.32}
\alias{ex15.32}
\docType{data}
\title{R Data set: ex15.32}
\description{
The \code{ex15.32} data frame has 7 rows and 2 columns.
}
\usage{data(ex15.32)}
\format{
  A data frame with 7 observations on the following 2 variables.
  \describe{
    \item{\code{Lateral}}{a numeric vector}
    \item{\code{Diagonal}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex15.32)
str(ex15.32)
}
\keyword{datasets}
