\name{ex15.35}
\alias{ex15.35}
\docType{data}
\title{R Data set: ex15.35}
\description{
The \code{ex15.35} data frame has 5 rows and 2 columns.
}
\usage{data(ex15.35)}
\format{
  A data frame with 5 observations on the following 2 variables.
  \describe{
    \item{\code{SIDS}}{a numeric vector}
    \item{\code{Control}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex15.35)
str(ex15.35)
}
\keyword{datasets}
