\name{ex16.06}
\alias{ex16.06}
\docType{data}
\title{R Data set: ex16.06}
\description{
The \code{ex16.06} data frame has 22 rows and 5 columns.
}
\usage{data(ex16.06)}
\format{
  A data frame with 22 observations on the following 5 variables.
  \describe{
    \item{\code{Obs.1}}{a numeric vector}
    \item{\code{Obs.2}}{a numeric vector}
    \item{\code{Obs.3}}{a numeric vector}
    \item{\code{Obs.4}}{a numeric vector}
    \item{\code{Obs.5}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex16.06)
str(ex16.06)
}
\keyword{datasets}
