\name{ex16.25}
\alias{ex16.25}
\docType{data}
\title{R Data set: ex16.25}
\description{
The \code{ex16.25} data frame has 22 rows and 3 columns.
}
\usage{data(ex16.25)}
\format{
  A data frame with 22 observations on the following 3 variables.
  \describe{
    \item{\code{C1}}{a factor with levels \code{1} \code{10} \code{11} \code{12} \code{13} \code{14} \code{15} \code{16} \code{17} \code{18} \code{19} \code{2} \code{20} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9} \code{C1} \code{Panel}}
    \item{\code{C2}}{a factor with levels \code{0.6} \code{0.8} \code{1} \code{Area Examined} \code{C2}}
    \item{\code{C3}}{a factor with levels \code{1} \code{10} \code{12} \code{2} \code{3} \code{4} \code{5} \code{6} \code{# Blemishes} \code{C3}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex16.25)
str(ex16.25)
}
\keyword{datasets}
