\name{ex16.41}
\alias{ex16.41}
\docType{data}
\title{R Data set: ex16.41}
\description{
The \code{ex16.41} data frame has 22 rows and 3 columns.
}
\usage{data(ex16.41)}
\format{
  A data frame with 22 observations on the following 3 variables.
  \describe{
    \item{\code{C1}}{a numeric vector}
    \item{\code{C2}}{a numeric vector}
    \item{\code{C3}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex16.41)
str(ex16.41)
}
\keyword{datasets}
