\name{ex16.43}
\alias{ex16.43}
\docType{data}
\title{R Data set: ex16.43}
\description{
The \code{ex16.43} data frame has 20 rows and 3 columns.
}
\usage{data(ex16.43)}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{Col1}}{a numeric vector}
    \item{\code{Col2}}{a numeric vector}
    \item{\code{Col3}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(ex16.43)
str(ex16.43)
}
\keyword{datasets}
