\name{xmp01.02}
\alias{xmp01.02}
\docType{data}
\title{R Data set: xmp01.02}
\description{
The \code{xmp01.02} data frame has 27 rows and 1 column.
}
\usage{data(xmp01.02)}
\format{
  A data frame with 27 observations on the following variable.
  \describe{
    \item{\code{C1}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp01.02)
str(xmp01.02)
}
\keyword{datasets}
