\name{xmp01.05}
\alias{xmp01.05}
\docType{data}
\title{R Data set: xmp01.05}
\description{
The \code{xmp01.05} data frame has 140 rows and 1 column.
}
\usage{data(xmp01.05)}
\format{
  A data frame with 140 observations on the following variable.
  \describe{
    \item{\code{bingePct}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp01.05)
str(xmp01.05)
}
\keyword{datasets}
