\name{xmp01.08}
\alias{xmp01.08}
\docType{data}
\title{R Data set: xmp01.08}
\description{
	a data set
}
\usage{data(xmp01.08)}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp01.08)
str(xmp01.08)
}
\keyword{datasets}
