\name{xmp01.16}
\alias{xmp01.16}
\docType{data}
\title{R Data set: xmp01.16}
\description{
	A data set
}
\usage{data(xmp01.16)}
\details{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp01.16)
str(xmp01.16)
}
\keyword{datasets}
