\name{xmp04.29}
\alias{xmp04.29}
\docType{data}
\title{R Data set: xmp04.29}
\description{
The \code{xmp04.29} data frame has 10 rows and 1 column.
}
\usage{data(xmp04.29)}
\format{
  A data frame with 10 observations on the following variable.
  \describe{
    \item{\code{meas.err}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp04.29)
str(xmp04.29)
}
\keyword{datasets}
