\name{xmp04.30}
\alias{xmp04.30}
\docType{data}
\title{R Data set: xmp04.30}
\description{
The \code{xmp04.30} data frame has 20 rows and 2 columns.
}
\usage{data(xmp04.30)}
\format{
  A data frame with 20 observations on the following 2 variables.
  \describe{
    \item{\code{Voltage}}{a numeric vector}
    \item{\code{z.percentile}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp04.30)
str(xmp04.30)
}
\keyword{datasets}
