\name{xmp07.15}
\alias{xmp07.15}
\docType{data}
\title{R Data set: xmp07.15}
\description{
The \code{xmp07.15} data frame has 17 rows and 1 column.
}
\usage{data(xmp07.15)}
\format{
  A data frame with 17 observations on the following variable.
  \describe{
    \item{\code{voltage}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp07.15)
str(xmp07.15)
}
\keyword{datasets}
