\name{xmp08.08}
\alias{xmp08.08}
\docType{data}
\title{R Data set: xmp08.08}
\description{
The \code{xmp08.08} data frame has 52 rows and 1 column.
}
\usage{data(xmp08.08)}
\format{
  A data frame with 52 observations on the following variable.
  \describe{
    \item{\code{DCP}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp08.08)
str(xmp08.08)
}
\keyword{datasets}
