\name{xmp08.09}
\alias{xmp08.09}
\docType{data}
\title{R Data set: xmp08.09}
\description{
The \code{xmp08.09} data frame has 5 rows and 1 column.
}
\usage{data(xmp08.09)}
\format{
  A data frame with 5 observations on the following variable.
  \describe{
    \item{\code{MAWL}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp08.09)
str(xmp08.09)
}
\keyword{datasets}
