\name{xmp09.06}
\alias{xmp09.06}
\docType{data}
\title{R Data set: xmp09.06}
\description{
The \code{xmp09.06} data frame has 2 rows and 4 columns.
}
\usage{data(xmp09.06)}
\format{
  A data frame with 2 observations on the following 4 variables.
  \describe{
    \item{\code{Fabric.Type}}{a factor with levels \code{Cotton} \code{Triacetate}}
    \item{\code{Sample.Size}}{a numeric vector}
    \item{\code{Sample.Mean}}{a numeric vector}
    \item{\code{Sample.Standard.Deviation}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp09.06)
str(xmp09.06)
}
\keyword{datasets}
