\name{xmp09.07}
\alias{xmp09.07}
\docType{data}
\title{R Data set: xmp09.07}
\description{
The \code{xmp09.07} data frame has 18 rows and 2 columns.
}
\usage{data(xmp09.07)}
\format{
  A data frame with 18 observations on the following 2 variables.
  \describe{
    \item{\code{strength}}{a numeric vector}
    \item{\code{type}}{a factor with levels \code{fused} \code{nofusion}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp09.07)
str(xmp09.07)
}
\keyword{datasets}
