\name{xmp09.08}
\alias{xmp09.08}
\docType{data}
\title{R Data set: xmp09.08}
\description{
The \code{xmp09.08} data frame has 6 rows and 2 columns.
}
\usage{data(xmp09.08)}
\format{
  A data frame with 6 observations on the following 2 variables.
  \describe{
    \item{\code{bottom}}{a numeric vector}
    \item{\code{surface}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp09.08)
str(xmp09.08)
}
\keyword{datasets}
