\name{xmp09.09}
\alias{xmp09.09}
\docType{data}
\title{R Data set: xmp09.09}
\description{
The \code{xmp09.09} data frame has 16 rows and 4 columns.
}
\usage{data(xmp09.09)}
\format{
  A data frame with 16 observations on the following 4 variables.
  \describe{
    \item{\code{Subject}}{a numeric vector}
    \item{\code{Before}}{a numeric vector}
    \item{\code{After}}{a numeric vector}
    \item{\code{Difference}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp09.09)
str(xmp09.09)
}
\keyword{datasets}
