\name{xmp09.10}
\alias{xmp09.10}
\docType{data}
\title{R Data set: xmp09.10}
\description{
The \code{xmp09.10} data frame has 13 rows and 2 columns.
}
\usage{data(xmp09.10)}
\format{
  A data frame with 13 observations on the following 2 variables.
  \describe{
    \item{\code{slide}}{a numeric vector}
    \item{\code{digital}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp09.10)
str(xmp09.10)
}
\keyword{datasets}
