\name{xmp10.01}
\alias{xmp10.01}
\docType{data}
\title{R Data set: xmp10.01}
\description{
The \code{xmp10.01} data frame has 24 rows and 2 columns.
}
\usage{data(xmp10.01)}
\format{
  A data frame with 24 observations on the following 2 variables.
  \describe{
    \item{\code{C1}}{a numeric vector}
    \item{\code{C2}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp10.01)
str(xmp10.01)
}
\keyword{datasets}
