\name{xmp10.03}
\alias{xmp10.03}
\docType{data}
\title{R Data set: xmp10.03}
\description{
The \code{xmp10.03} data frame has 15 rows and 2 columns.
}
\usage{data(xmp10.03)}
\format{
  A data frame with 15 observations on the following 2 variables.
  \describe{
    \item{\code{Soiling}}{a numeric vector}
    \item{\code{Mixture}}{a factor with levels \code{1} \code{2} \code{3}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp10.03)
str(xmp10.03)
}
\keyword{datasets}
