\name{xmp10.05}
\alias{xmp10.05}
\docType{data}
\title{R Data set: xmp10.05}
\description{
The \code{xmp10.05} data frame has 20 rows and 2 columns.
}
\usage{data(xmp10.05)}
\format{
  A data frame with 20 observations on the following 2 variables.
  \describe{
    \item{\code{REMtime}}{a numeric vector}
    \item{\code{ethanol}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp10.05)
str(xmp10.05)
}
\keyword{datasets}
