\name{xmp10.08}
\alias{xmp10.08}
\docType{data}
\title{R Data set: xmp10.08}
\description{
The \code{xmp10.08} data frame has 22 rows and 2 columns.
}
\usage{data(xmp10.08)}
\format{
  A data frame with 22 observations on the following 2 variables.
  \describe{
    \item{\code{elastic}}{a numeric vector}
    \item{\code{type}}{a factor with levels \code{Die} \code{Permanent} \code{Plaster}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp10.08)
str(xmp10.08)
}
\keyword{datasets}
