\name{xmp10.10}
\alias{xmp10.10}
\docType{data}
\title{R Data set: xmp10.10}
\description{
The \code{xmp10.10} data frame has 18 rows and 2 columns.
}
\usage{data(xmp10.10)}
\format{
  A data frame with 18 observations on the following 2 variables.
  \describe{
    \item{\code{travel}}{a numeric vector}
    \item{\code{Rail}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp10.10)
str(xmp10.10)
}
\keyword{datasets}
