\name{xmp11.01}
\alias{xmp11.01}
\docType{data}
\title{R Data set: xmp11.01}
\description{
The \code{xmp11.01} data frame has 12 rows and 3 columns.
}
\usage{data(xmp11.01)}
\format{
  A data frame with 12 observations on the following 3 variables.
  \describe{
    \item{\code{strength}}{a numeric vector}
    \item{\code{brand}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{treatment}}{a factor with levels \code{1} \code{2} \code{3} \code{4}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp11.01)
str(xmp11.01)
}
\keyword{datasets}
