\name{xmp11.05}
\alias{xmp11.05}
\docType{data}
\title{R Data set: xmp11.05}
\description{
The \code{xmp11.05} data frame has 20 rows and 3 columns.
}
\usage{data(xmp11.05)}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{power}}{a numeric vector}
    \item{\code{humid}}{an ordered factor with levels \code{1} < \code{2} < \code{3} < \code{4}}
    \item{\code{brand}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp11.05)
str(xmp11.05)
}
\keyword{datasets}
