\name{xmp11.06}
\alias{xmp11.06}
\docType{data}
\title{R Data set: xmp11.06}
\description{
The \code{xmp11.06} data frame has 24 rows and 3 columns.
}
\usage{data(xmp11.06)}
\format{
  A data frame with 24 observations on the following 3 variables.
  \describe{
    \item{\code{Resp}}{a numeric vector}
    \item{\code{Stimulus}}{a factor with levels \code{L1} \code{L2} \code{T} \code{L1+L2} \code{L1+T} \code{L2+T}}
    \item{\code{Subject}}{a factor with levels \code{1} \code{2} \code{3} \code{4}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp11.06)
str(xmp11.06)
}
\keyword{datasets}
