\name{xmp11.07}
\alias{xmp11.07}
\docType{data}
\title{R Data set: xmp11.07}
\description{
The \code{xmp11.07} data frame has 36 rows and 3 columns.
}
\usage{data(xmp11.07)}
\format{
  A data frame with 36 observations on the following 3 variables.
  \describe{
    \item{\code{Yield}}{a numeric vector}
    \item{\code{Variety}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{Density}}{a factor with levels \code{1} \code{2} \code{3} \code{4}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp11.07)
str(xmp11.07)
}
\keyword{datasets}
