\name{xmp11.10}
\alias{xmp11.10}
\docType{data}
\title{R Data set: xmp11.10}
\description{
The \code{xmp11.10} data frame has 96 rows and 4 columns.
}
\usage{data(xmp11.10)}
\format{
  A data frame with 96 observations on the following 4 variables.
  \describe{
    \item{\code{Tempr}}{a numeric vector}
    \item{\code{Period}}{a numeric vector}
    \item{\code{Strain}}{a numeric vector}
    \item{\code{Coat}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp11.10)
str(xmp11.10)
}
\keyword{datasets}
