\name{xmp11.11}
\alias{xmp11.11}
\docType{data}
\title{R Data set: xmp11.11}
\description{
The \code{xmp11.11} data frame has 36 rows and 4 columns.
}
\usage{data(xmp11.11)}
\format{
  A data frame with 36 observations on the following 4 variables.
  \describe{
    \item{\code{abrasion}}{a numeric vector}
    \item{\code{row}}{an ordered factor with levels \code{1} < \code{2} < \code{3} < \code{4} < \code{5} < \code{6}}
    \item{\code{column}}{an ordered factor with levels \code{1} < \code{2} < \code{3} < \code{4} < \code{5} < \code{6}}
    \item{\code{humidity}}{a factor with levels \code{25 percent} \code{37 percent} \code{50 percent} \code{62 percent} \code{75 percent} \code{87 percent}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp11.11)
str(xmp11.11)
}
\keyword{datasets}
