\name{xmp11.12}
\alias{xmp11.12}
\docType{data}
\title{R Data set: xmp11.12}
\description{
The \code{xmp11.12} data frame has 8 rows and 4 columns.
}
\usage{data(xmp11.12)}
\format{
  A data frame with 8 observations on the following 4 variables.
  \describe{
    \item{\code{Age}}{a numeric vector}
    \item{\code{Temperature}}{a numeric vector}
    \item{\code{Soil.1}}{a numeric vector}
    \item{\code{Soil.2}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp11.12)
str(xmp11.12)
}
\keyword{datasets}
