\name{xmp12.01}
\alias{xmp12.01}
\docType{data}
\title{R Data set: xmp12.01}
\description{
The \code{xmp12.01} data frame has 30 rows and 3 columns.
}
\usage{data(xmp12.01)}
\format{
  A data frame with 30 observations on the following 3 variables.
  \describe{
    \item{\code{Obs}}{a numeric vector}
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp12.01)
str(xmp12.01)
}
\keyword{datasets}
