\name{xmp12.02}
\alias{xmp12.02}
\docType{data}
\title{R Data set: xmp12.02}
\description{
The \code{xmp12.02} data frame has 19 rows and 2 columns.
}
\usage{data(xmp12.02)}
\format{
  A data frame with 19 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp12.02)
str(xmp12.02)
}
\keyword{datasets}
