\name{xmp12.06}
\alias{xmp12.06}
\docType{data}
\title{R Data set: xmp12.06}
\description{
The \code{xmp12.06} data frame has 20 rows and 2 columns.
}
\usage{data(xmp12.06)}
\format{
  A data frame with 20 observations on the following 2 variables.
  \describe{
    \item{\code{moistcon}}{a numeric vector}
    \item{\code{filtrate}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp12.06)
str(xmp12.06)
}
\keyword{datasets}
