\name{xmp13.01}
\alias{xmp13.01}
\docType{data}
\title{R Data set: xmp13.01}
\description{
The \code{xmp13.01} data frame has 14 rows and 2 columns.
}
\usage{data(xmp13.01)}
\format{
  A data frame with 14 observations on the following 2 variables.
  \describe{
    \item{\code{xi}}{a numeric vector}
    \item{\code{yi}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp13.01)
str(xmp13.01)
}
\keyword{datasets}
