\name{xmp13.03}
\alias{xmp13.03}
\docType{data}
\title{R Data set: xmp13.03}
\description{
The \code{xmp13.03} data frame has 12 rows and 2 columns.
}
\usage{data(xmp13.03)}
\format{
  A data frame with 12 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp13.03)
str(xmp13.03)
}
\keyword{datasets}
