\name{xmp13.06}
\alias{xmp13.06}
\docType{data}
\title{R Data set: xmp13.06}
\description{
The \code{xmp13.06} data frame has 24 rows and 2 columns.
}
\usage{data(xmp13.06)}
\format{
  A data frame with 24 observations on the following 2 variables.
  \describe{
    \item{\code{Temperature}}{a numeric vector}
    \item{\code{Failure}}{a factor with levels \code{N} \code{Y}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp13.06)
str(xmp13.06)
}
\keyword{datasets}
