\name{xmp13.11}
\alias{xmp13.11}
\docType{data}
\title{R Data set: xmp13.11}
\description{
The \code{xmp13.11} data frame has 30 rows and 6 columns.
}
\usage{data(xmp13.11)}
\format{
  A data frame with 30 observations on the following 6 variables.
  \describe{
    \item{\code{Observation}}{a numeric vector}
    \item{\code{Force}}{a numeric vector}
    \item{\code{Power}}{a numeric vector}
    \item{\code{Temperature}}{a numeric vector}
    \item{\code{Time}}{a numeric vector}
    \item{\code{Strength}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp13.11)
str(xmp13.11)
}
\keyword{datasets}
