\name{xmp13.13}
\alias{xmp13.13}
\docType{data}
\title{R Data set: xmp13.13}
\description{
The \code{xmp13.13} data frame has 9 rows and 5 columns.
}
\usage{data(xmp13.13)}
\format{
  A data frame with 9 observations on the following 5 variables.
  \describe{
    \item{\code{x1}}{a numeric vector}
    \item{\code{x2}}{a numeric vector}
    \item{\code{x1x2}}{a numeric vector}
    \item{\code{X28}}{a numeric vector}
    \item{\code{Absorbability}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp13.13)
str(xmp13.13)
}
\keyword{datasets}
