\name{xmp13.15}
\alias{xmp13.15}
\docType{data}
\title{R Data set: xmp13.15}
\description{
The \code{xmp13.15} data frame has 13 rows and 3 columns.
}
\usage{data(xmp13.15)}
\format{
  A data frame with 13 observations on the following 3 variables.
  \describe{
    \item{\code{Iron.x1}}{a numeric vector}
    \item{\code{Aluminum.x2}}{a numeric vector}
    \item{\code{Adsorption.y}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp13.15)
str(xmp13.15)
}
\keyword{datasets}
