\name{xmp13.18}
\alias{xmp13.18}
\docType{data}
\title{R Data set: xmp13.18}
\description{
The \code{xmp13.18} data frame has 27 rows and 3 columns.
}
\usage{data(xmp13.18)}
\format{
  A data frame with 27 observations on the following 3 variables.
  \describe{
    \item{\code{s}}{a numeric vector}
    \item{\code{l.1000.s}}{a numeric vector}
    \item{\code{w}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp13.18)
str(xmp13.18)
}
\keyword{datasets}
