\name{xmp13.19}
\alias{xmp13.19}
\docType{data}
\title{R Data set: xmp13.19}
\description{
The \code{xmp13.19} data frame has 31 rows and 5 columns.
}
\usage{data(xmp13.19)}
\format{
  A data frame with 31 observations on the following 5 variables.
  \describe{
    \item{\code{y}}{a numeric vector}
    \item{\code{x1}}{a numeric vector}
    \item{\code{x2}}{a numeric vector}
    \item{\code{x.1}}{a numeric vector}
    \item{\code{x.2}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp13.19)
str(xmp13.19)
}
\keyword{datasets}
