\name{xmp13.22}
\alias{xmp13.22}
\docType{data}
\title{R Data set: xmp13.22}
\description{
The \code{xmp13.22} data frame has 10 rows and 3 columns.
}
\usage{data(xmp13.22)}
\format{
  A data frame with 10 observations on the following 3 variables.
  \describe{
    \item{\code{Strength}}{a numeric vector}
    \item{\code{Sp.grav}}{a numeric vector}
    \item{\code{Moisture}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp13.22)
str(xmp13.22)
}
\keyword{datasets}
