\name{xmp14.03}
\alias{xmp14.03}
\docType{data}
\title{R Data set: xmp14.03}
\description{
The \code{xmp14.03} data frame has 24 rows and 1 column.
}
\usage{data(xmp14.03)}
\format{
  A data frame with 24 observations on the following variable.
  \describe{
    \item{\code{onset}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp14.03)
str(xmp14.03)
}
\keyword{datasets}
