\name{xmp14.13}
\alias{xmp14.13}
\docType{data}
\title{R Data set: xmp14.13}
\description{
The \code{xmp14.13} data frame has 4 rows and 7 columns.
}
\usage{data(xmp14.13)}
\format{
  A data frame with 4 observations on the following 7 variables.
  \describe{
    \item{\code{Production.Line}}{a numeric vector}
    \item{\code{Blemish}}{a numeric vector}
    \item{\code{Crack}}{a numeric vector}
    \item{\code{Location}}{a numeric vector}
    \item{\code{Missing}}{a numeric vector}
    \item{\code{Other}}{a numeric vector}
    \item{\code{Size}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp14.13)
str(xmp14.13)
}
\keyword{datasets}
