\name{xmp14.14}
\alias{xmp14.14}
\docType{data}
\title{R Data set: xmp14.14}
\description{
The \code{xmp14.14} data frame has 3 rows and 3 columns.
}
\usage{data(xmp14.14)}
\format{
  A data frame with 3 observations on the following 3 variables.
  \describe{
    \item{\code{Substand}}{a numeric vector}
    \item{\code{Standard}}{a numeric vector}
    \item{\code{Modern}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp14.14)
str(xmp14.14)
}
\keyword{datasets}
