\name{xmp15.01}
\alias{xmp15.01}
\docType{data}
\title{R Data set: xmp15.01}
\description{
The \code{xmp15.01} data frame has 15 rows and 1 column.
}
\usage{data(xmp15.01)}
\format{
  A data frame with 15 observations on the following variable.
  \describe{
    \item{\code{C1}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp15.01)
str(xmp15.01)
}
\keyword{datasets}
