\name{xmp15.02}
\alias{xmp15.02}
\docType{data}
\title{R Data set: xmp15.02}
\description{
The \code{xmp15.02} data frame has 8 rows and 5 columns.
}
\usage{data(xmp15.02)}
\format{
  A data frame with 8 observations on the following 5 variables.
  \describe{
    \item{\code{Log}}{a numeric vector}
    \item{\code{Solvent.1}}{a numeric vector}
    \item{\code{Solvent.2}}{a numeric vector}
    \item{\code{Difference}}{a numeric vector}
    \item{\code{Signed.rank}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp15.02)
str(xmp15.02)
}
\keyword{datasets}
