\name{xmp15.03}
\alias{xmp15.03}
\docType{data}
\title{R Data set: xmp15.03}
\description{
The \code{xmp15.03} data frame has 25 rows and 2 columns.
}
\usage{data(xmp15.03)}
\format{
  A data frame with 25 observations on the following 2 variables.
  \describe{
    \item{\code{xi}}{a numeric vector}
    \item{\code{Signed.Rank}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp15.03)
str(xmp15.03)
}
\keyword{datasets}
