\name{xmp15.09}
\alias{xmp15.09}
\docType{data}
\title{R Data set: xmp15.09}
\description{
The \code{xmp15.09} data frame has 35 rows and 2 columns.
}
\usage{data(xmp15.09)}
\format{
  A data frame with 35 observations on the following 2 variables.
  \describe{
    \item{\code{C1}}{a numeric vector}
    \item{\code{C2}}{a factor with levels \code{10\"} \code{12\"} \code{4\"} \code{6\"} \code{8\"}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp15.09)
str(xmp15.09)
}
\keyword{datasets}
