\name{xmp15.10}
\alias{xmp15.10}
\docType{data}
\title{R Data set: xmp15.10}
\description{
The \code{xmp15.10} data frame has 8 rows and 4 columns.
}
\usage{data(xmp15.10)}
\format{
  A data frame with 8 observations on the following 4 variables.
  \describe{
    \item{\code{Fear}}{a numeric vector}
    \item{\code{Happiness}}{a numeric vector}
    \item{\code{Depression}}{a numeric vector}
    \item{\code{Calmness}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp15.10)
str(xmp15.10)
}
\keyword{datasets}
