\name{xmp16.01}
\alias{xmp16.01}
\docType{data}
\title{R Data set: xmp16.01}
\description{
The \code{xmp16.01} data frame has 25 rows and 3 columns.
}
\usage{data(xmp16.01)}
\format{
  A data frame with 25 observations on the following 3 variables.
  \describe{
    \item{\code{Visc1}}{a numeric vector}
    \item{\code{Visc2}}{a numeric vector}
    \item{\code{Visc3}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp16.01)
str(xmp16.01)
}
\keyword{datasets}
