\name{xmp16.06}
\alias{xmp16.06}
\docType{data}
\title{R Data set: xmp16.06}
\description{
The \code{xmp16.06} data frame has 27 rows and 26 columns.
}
\usage{data(xmp16.06)}
\format{
  A data frame with 27 observations on the following 26 variables.
  \describe{
    \item{\code{C1}}{a factor with levels \code{1} \code{10} \code{11} \code{12} \code{13} \code{14} \code{15} \code{16} \code{17} \code{18} \code{19} \code{2} \code{20} \code{21} \code{22} \code{23} \code{24} \code{25} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9} \code{Day (i)}}
    \item{\code{C2}}{a numeric vector}
    \item{\code{C3}}{a numeric vector}
    \item{\code{C4}}{a numeric vector}
    \item{\code{C5}}{a numeric vector}
    \item{\code{C6}}{a numeric vector}
    \item{\code{C7}}{a numeric vector}
    \item{\code{C8}}{a numeric vector}
    \item{\code{C9}}{a numeric vector}
    \item{\code{C10}}{a numeric vector}
    \item{\code{C11}}{a numeric vector}
    \item{\code{C12}}{a numeric vector}
    \item{\code{C13}}{a numeric vector}
    \item{\code{C14}}{a numeric vector}
    \item{\code{C15}}{a numeric vector}
    \item{\code{C16}}{a numeric vector}
    \item{\code{C17}}{a numeric vector}
    \item{\code{C18}}{a numeric vector}
    \item{\code{C19}}{a numeric vector}
    \item{\code{C20}}{a numeric vector}
    \item{\code{C21}}{a numeric vector}
    \item{\code{C22}}{a numeric vector}
    \item{\code{C23}}{a numeric vector}
    \item{\code{C24}}{a numeric vector}
    \item{\code{C25}}{a numeric vector}
    \item{\code{C26}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp16.06)
str(xmp16.06)
}
\keyword{datasets}
