\name{xmp16.08}
\alias{xmp16.08}
\docType{data}
\title{R Data set: xmp16.08}
\description{
The \code{xmp16.08} data frame has 16 rows and 4 columns.
}
\usage{data(xmp16.08)}
\format{
  A data frame with 16 observations on the following 4 variables.
  \describe{
    \item{\code{Obs.1}}{a numeric vector}
    \item{\code{Obs.2}}{a numeric vector}
    \item{\code{Obs.3}}{a numeric vector}
    \item{\code{Obs.4}}{a numeric vector}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp16.08)
str(xmp16.08)
}
\keyword{datasets}
