\name{xmp16.09}
\alias{xmp16.09}
\docType{data}
\title{R Data set: xmp16.09}
\description{
The \code{xmp16.09} data frame has 18 rows and 6 columns.
}
\usage{data(xmp16.09)}
\format{
  A data frame with 18 observations on the following 6 variables.
  \describe{
    \item{\code{C1}}{a factor with levels \code{1} \code{10} \code{11} \code{12} \code{13} \code{14} \code{15} \code{16} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9} \code{C1} \code{Sample #}}
    \item{\code{C2}}{a factor with levels \code{39.65} \code{39.72} \code{39.76} \code{39.84} \code{39.98} \code{40.06} \code{40.2} \code{40.23} \code{40.32} \code{40.34} \code{40.4} \code{40.41} \code{40.42} \code{40.49} \code{40.61} \code{C2} \code{xwl}}
    \item{\code{C3}}{a factor with levels \code{0.05} \code{0.08} \code{-0.09} \code{-0.17} \code{0.17} \code{0.19} \code{0.25} \code{0.26} \code{0.27} \code{-0.31} \code{0.34} \code{-0.39} \code{-0.43} \code{0.46} \code{-0.5} \code{C3} \code{xwl - 40.15}}
    \item{\code{C4}}{a factor with levels \code{0} \code{0.01} \code{0.05} \code{0.1} \code{0.12} \code{0.26} \code{0.27} \code{0.43} \code{0.46} \code{0.71} \code{1} \code{1.08} \code{1.17} \code{1.27} \code{C4} \code{dl}}
    \item{\code{C5}}{a factor with levels \code{-0.01} \code{-0.09} \code{-0.13} \code{0.13} \code{-0.2} \code{0.21} \code{0.35} \code{0.38} \code{0.47} \code{0.49} \code{0.55} \code{0.56} \code{0.57} \code{0.64} \code{0.76} \code{C5} \code{xwl - 39.85}}
    \item{\code{C6}}{a factor with levels \code{0} \code{0.01} \code{0.09} \code{0.13} \code{0.29} \code{C6} \code{el}}
  }
}
\details{
Consult the web site \url{http://www.thomsonedu.com/statistics/devore} for additional online resources that are available for this book.
}
\source{
Devore, J. L. (2008) \emph{Probability and Statistics for Engineering and the Sciences (7th Edition)},  ISBN-10: 0495382175   ISBN-13: 9780495382171
}
\examples{
data(xmp16.09)
str(xmp16.09)
}
\keyword{datasets}
